tags: Factor get_tags out of the _get_tags endpoint
Gettings a nicely-parsed list of tags is useful internally, so pull
the tag-listing logic out into it's own function.  The decorated
endpoint behind GET /v1/repositories/<path:repository>/tags is *not*
useful internally, so prefix that function with an underscore.

To convert the get_tags iterator to a dictionary, I've used a dict
comprehension [1], which landed in Python 2.7 and 3.0.  I couldn't
find any docs on docker-registry's minimal supported Python version,
so I imagine compatibility with ancient Pythons is not high on the
agenda.

[1]: http://www.python.org/dev/peps/pep-0274/