tags: Restore delete_repository URL to the spec value
The rename from

  DELETE /v1/repositories/<namespace>/<path:repository>/

(which matches the spec [1,2]) to

  DELETE /v1/repositories/<namespace>/<path:repository>/tags

(not in the spec) happened in 1026a87 (renaming DELETE call on a
repos, 2013-06-26).  The idea was that if the registry is *not* in
standalone mode and the index-side handling was disabled [2], we'd
still want an endpoint that implemented the registry-side cleanup
[1,3].

This commit restores the original URL (but leaves the function to the
tags module to work regardless of standalone mode), triggers
tag-deleted signals for each tag, removes the remaining repository
data en masse (which handles the index-side removal as well), and
triggers the repository-deleted signal.  It doesn't remove referenced
images, because those images might be required by other repositories.
We'd want some sort of refcounting (number of repositories/tags
requiring this image) to decide when image-removal was allowed.

[1]: http://docs.docker.io/en/latest/reference/api/registry_api/#delete--v1-repositories-%28namespace%29-%28repository%29-
[2]: http://docs.docker.io/en/latest/reference/api/index_api/#delete--v1-repositories-%28namespace%29-%28repo_name%29-
[3]: https://github.com/dotcloud/docker-registry/pull/268#issuecomment-36470080