Skip test_list_directory_with_subdir in S3 driver.
The base use case for test_list_directory_with_subdir is to test that
the local database is recreated correctly.  I tested this with the
current codebase and it is indeed working.

The reason this fails is that the lookup/list mocked functions are not
close enough from reality.  For more details see GH #596.