Add before_put_image_json signal
Add before_put_image_json signal that allows receivers to short-circuit
the put if they return any non-None value. This can be useful for things
such as enforcing quotas, e.g. by having a custom extension connect to
this signal, check the layer's Size, and returning something to the
effect that the put was denied due to quota.

Signed-off-by: Andy Goldstein <agoldste@redhat.com>