Use gevent to parallelize storage operations in get_tags
This is a band-aid for #614

For context, hosted from a GCE instance with GCS being the backing
store, With this patch, we are able to 'docker pull' 859 tags without a
timeout.  Prior to this patch, we were *unable* to pull anything from a
repository with 200 tags.  This is the extent of my investigation thus
far.  Unfortunately, this change is probably just a constant factor
reduction in run-time as total network round-trips still grows linearly
with tag count.